First attempt at velocity calculation
Accelerometers in mobile devices may not be accurate, there will be a lot of noise.
Tried adding reasonable filters to reduce noise.
Currently can't be sure if the calculated velocity(ms) is accurate.
Consider using GPS to get better accuracy, but might not be necessary if  accelerometer accuracy can be improved.